Create ConsoleManager class, refactor SaladBuilder, add test for SaladBuilder, redid SearchingVegetablesTest.

+++ tasks/VitaliHardzeyenka/ChiefCooker/src/com/epam/training/ChiefCooker/Constants/SolenaceaeConstants.java
@@ -1,9 +1,9 @@
 package com.epam.training.ChiefCooker.Constants;
 
 public class SolenaceaeConstants {
-	public final static double COEFFICIENT_WEIGHT_SEEDS_NORMAL_GOOD = (double) 0.15;
-	public final static double COEFFICIENT_WEIGHT_SEEDS_BAD = (double) 0.05;
-	public final static double COEFFICIENT_WEIGHT_PULP_BAD = (double) 0.4;
-	public final static double COEFFICIENT_WEIGHT_PULP_NORMAL = (double) 0.2;
-	public final static double COEFFICIENT_WEIGHT_PULP_GOOD = (double) 0.0;
+	public final static double COEFFICIENT_WEIGHT_SEEDS_NORMAL_GOOD = 0.15d;
+	public final static double COEFFICIENT_WEIGHT_SEEDS_BAD = 0.05d;
+	public final static double COEFFICIENT_WEIGHT_PULP_BAD = 0.4d;
+	public final static double COEFFICIENT_WEIGHT_PULP_NORMAL = 0.2d;
+	public final static double COEFFICIENT_WEIGHT_PULP_GOOD = 0.0d;
 }
